Separate the admin dashboard layout
